[release/6.0] Don't create multiple large files at the same time (#63032)
* Move NoInt32OverflowInTheBufferingLogic to OuterLoop and address pending feedback (#60606)

* Don't create multiple large files at the same time (#62519)

* move existing large file tests into a separate type (no code changes)

* don't run the large file tests in parallel

* use FileOptions.DeleteOnClose to ensure that each test removes it's own file

use single large file to test File.ReadAllBytes and ile.ReadAllBytesAsync for both limits

* Fix DisableParallelization build issue

* Apply suggestions from code review


* Update src/libraries/System.IO.FileSystem/tests/LargeFileTests.cs


* Notepad with uppercase (#62487)
